Fixing Build Bug

+++ SortingVisualizerCSharp/Painter.cs
@@ -9,7 +9,7 @@ namespace SortingVisualizerCSharp
   {
     public static void PaintLine(double value, Graphics g, int posX, int posY, int w)
     {
-      g.FillRectangle(new SolidBrush(Color.Black), posX, 0, w, (int)value);
+      g.FillRectangle(new SolidBrush(Color.Black), posX, 0, w, (float)value);
       //g.FillRectangle(new SolidBrush(Color.Black), 100, 400, 100, 100);
     }
   }
